Trial 5: Trial Algorithm
Keep track of the value of the state counter, if object is further than safe distance add to it, if not reduce from it. Values are then mapped to the six states and base on each state the servo will move to a specific angle and LED turn to a specific colour.